src/requirements.txt
asgiref==3.9.1
crispy-bootstrap4==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==25.1
django-localflavor==5.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-stdnum==2.1
sqlparse==0.5.3

